pyproject.toml
[tool.poetry]
name = "pyopenrefine"
version = "0.1.0"
description = ""
authors = ["Francesco Frassinelli <francesco.frassinelli@nina.no>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.group.dev.dependencies]
pytest-container = "^0.2.0"
pytest-xdist = "^3.3.1"

[tool.black]

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/black_compatibility.html
profile = "black"

[tool.ruff]
fix = true
ignore = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
